style: Modal Component style 구현
