+++ src/components/modal/style.scss
@@ -1,0 +1,30 @@
+.modal {
+  display: flex;
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  backdrop-filter: blur(5px);
+  z-index: 2;
+
+  .modal-inner {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin: 0 auto;
+    padding: 2rem;
+    max-width: 35rem;
+    min-width: 25rem;
+    min-height: 20rem;
+    border-radius: 0.8rem;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    box-shadow: 2px 2px 10px 4px #d2d2d2;
+    background: white;
+    z-index: 3;
+  }
+}
