+++ src/components/receipt/lottery-ball/index.js
