+++ src/components/receipt/lottery-ball/style.scss
