+++ src/components/receipt/purchase-number-item/index.js
