+++ src/components/receipt/purchase-number-item/style.scss
