+++ src/components/receipt/purchse-number-list/style.scss
